pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forum-analyzer"
version = "0.1.0"
description = "Shopify Developer Forum Analyzer"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Forum Analyzer Team"}
]
keywords = ["shopify", "forum", "analyzer", "discourse"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.1",
    "rich>=13.5.0",
    "structlog>=23.1.0",
    "click>=8.1.0",
    "tenacity>=8.2.0",
]

[project.scripts]
forum-analyzer = "forum_analyzer.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/shopify-dev-forum-analyzer"
Repository = "https://github.com/yourusername/shopify-dev-forum-analyzer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 79
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"